Refactor: Modular SYNNEX-to-Shopify Sync Pipeline
📝 PR Description:
This PR refactors the legacy monolithic synnex.js file into a fully modular, scalable sync pipeline. The new architecture separates responsibilities into clear, testable components while maintaining full backward compatibility.

✅ Key Changes:
Modularized pipeline:

scraper.js – Handles Playwright-based data scraping

apiFetcher.js – Pulls pricing and inventory from SYNNEX XML API

mappings.js – Maps product + metafields for Shopify

sync.js – Coordinates all logic to push products

New Job Runner:

syncSkusFromFile.js – Reads SKUs from a JSON file and triggers sync per item

Routes updated:

POST /api/synnex/sync/lenovo – Triggers full sync from batch file

Logs: All screenshots, XML dumps, and failures are saved in /logs

🧪 Tested:
 Synced 2 real SKUs from SYNNEX to Shopify

 Verified all metafields created correctly

 Scraped data included: image, specs, UPC, UNSPSC, warranty

 Failure handling: session expiry, missing fields, and API errors all logged

⚠️ Notes:
services/synnex/synnex.js is retained as legacy fallback for now

Cookie session stored in /data/cookies.json

No crons or automation yet — all routes are still manually triggered

📦 Future Enhancements:
Sync result logging to MongoDB

Admin UI for batch uploads + status tracking

Cron job for nightly sync

Queued retries for failed SKUs
