+++ routes/synnex.routes.js
@@ -1,6 +1,7 @@
 import express from "express";
-import { syncSynnexToShopify } from "../services/synnex/synnex.js";
-import { syncLenovoSKUsFromFile } from "../services/synnex/sync.js";
+import { syncSynnexToShopify } from "../services/synnex/sync.js";
+
+import { syncLenovoSKUsFromFile } from "../jobs/syncLenovoBatch.js";
 
 const router = express.Router();
 
@@ -27,6 +28,29 @@ router.post("/sync/lenovo", async (req, res) => {
   } catch (err) {
     console.error("❌ Sync error:", err.message);
     res.status(500).json({ error: "Failed to sync Lenovo SKUs" });
+  }
+});
+
+router.get("/test-sync", async (req, res) => {
+  const sku = req.query.sku || "7984359"; // default test SKU from your JSON
+
+  try {
+    const result = await syncSynnexToShopify(sku);
+    if (!result) {
+      return res.status(404).json({ error: "Product not synced (no data)" });
+    }
+
+    res.json({
+      message: `✅ Test sync successful for SKU: ${sku}`,
+      productId: result.id,
+      title: result.title,
+    });
+  } catch (err) {
+    console.error("❌ Sync error:", err.message);
+    res.status(500).json({
+      error: "Sync failed",
+      details: err.message,
+    });
   }
 });
